Códigos Bitwise + revisão arquivos de texto

+++ prova 3/lista 3 - Bitwise/exe1.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
-unsigned int matrizParaBits(int M[4][4]) {
-    unsigned int resultado=0;
+unsigned short int matrizParaBits(int M[4][4]){
+    unsigned short int resultado=0;
 
     for(int i=0;i<4;i++){
         for(int j=0;j<4;j++){
-            resultado= resultado<<1;
+            resultado=resultado<<1;
             resultado|=M[i][j];
         }
     }
